third_party/blink/web_tests/external/wpt/css/css-mixins/shadow-dom-expected.txt

This is a testharness.js-based test.
[FAIL] Style in shadow DOM should have access to outside non-adopted mixins
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
